Create @App Odd and Even Jumps
We are given an input string, We are also given two numbers: oddJump and evenJump. We pass through the odd letters from left to right and we sum their ASCII codes and aggregate the sum to the result (initially starting form 0). Through a step of oddJump we multiply the current result by the ASCII code of the current letter instead of adding it to the result. Finally we print the result as hexadecimal number. We do the same for the even letters with a step of evenJump. For our example, let's assume oddJump = 3 and evenJump = 2. 

•	odd result = ( ( (+ 115 's' + 102 'f') * 119 'w' + 114 'r' + 117 'u' ) * 105 'i' + 101 'e' + 115 's' ) * 116 't' = 317362776 = 12EA9258 (hex)
•	even result = ( ( + 111 'o' * 116 't' + 97 'a' ) * 101 'e' + 110 'n' ) * 118 'v' + 114 'r' ) * 105 'i' + 121 'y' = 16235657461 = 3C7B878F5 (hex)